requirements.txt
aiohttp==3.9.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pytz==2024.1
tqdm==4.66.1
rich==13.7.0
requests==2.31.0
beautifulsoup4==4.12.3
html2text==2020.1.16
